integration-tests: Adapt LP polkadot tests to use the generic framework
