+++ runtime/integration-tests/src/liquidity_pools/pallet/development/tests/liquidity_pools/add_allow_upgrade.rs
@@ -53,7 +53,6 @@ use xcm::{
 	v3::{Junction, Junctions},
 	VersionedMultiLocation,
 };
-use xcm_simulator::TestExt;
 
 use crate::{
 	chain::{
