+++ runtime/integration-tests/src/liquidity_pools/pallet/development/tests/liquidity_pools/foreign_investments.rs
@@ -64,7 +64,6 @@ use sp_runtime::{
 	WeakBoundedVec,
 };
 use tokio::runtime::Handle;
-use xcm_simulator::TestExt;
 
 use crate::{
 	chain::centrifuge::{
