--- runtime/integration-tests/src/liquidity_pools/pallet/polkadot/mod.rs
@@ -1,3 +1,0 @@
-mod setup;
-mod test_net;
-mod tests;
